Add support for dedicated physical address contact
If a physical/company address is different to billing
we store it as a separate contact on the customer.

Also some minor changes to email address handling for the
company partner.

Minor test/tox changes.

Changes project_name handling such that a task update
which changes the company name will correctly reflect
in the project name.

Change-Id: Iaa64613a746afbab64a85dde1fb5d94c445b3ccd